+++ src/test/java/ru/aston/popov_am/task9/OneTableForEveryoneClassWithUnions/UnionTableTest.java
@@ -1,34 +1,27 @@
-package ru.aston.popov_am.task9.OneTableForHierarchyTest;
+package ru.aston.popov_am.task9.OneTableForEveryoneClassWithUnions;
 
 import org.hibernate.Session;
 import org.hibernate.SessionFactory;
 import org.hibernate.cfg.Configuration;
-import org.junit.jupiter.api.AfterEach;
-import org.junit.jupiter.api.BeforeAll;
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Test;
-import ru.aston.popov_am.task9.OneTableForHierarchy.Cat;
-import ru.aston.popov_am.task9.OneTableForHierarchy.Fish;
-import ru.aston.popov_am.task9.OneTableForHierarchy.Pet;
+import org.junit.jupiter.api.*;
 
 import java.util.List;
 
-
-public class OneTableForHierarchyTestTable {
+public class UnionTableTest {
     private static SessionFactory sessionFactory;
     private Session session;
 
+    private Dog dog;
     private Cat cat;
-    private Fish fish;
 
     @BeforeAll
     public static void init() {
         try {
             Configuration configuration = new Configuration();
             configuration.configure("hibernate.cfg.xml");
-            configuration.addAnnotatedClass(Pet.class);
+            configuration.addAnnotatedClass(Animal.class);
+            configuration.addAnnotatedClass(Dog.class);
             configuration.addAnnotatedClass(Cat.class);
-            configuration.addAnnotatedClass(Fish.class);
             sessionFactory = configuration.buildSessionFactory();
         } catch (Exception e) {
             System.out.println(e.getMessage());
@@ -40,8 +33,8 @@ public class OneTableForHierarchyTestTable {
         try {
             session = sessionFactory.openSession();
             session.beginTransaction();
-            cat = Cat.builder().nickname("Мурзик").breed("Дворовая").color("Белая").build();
-            fish = Fish.builder().nickname("Тритон").waterTemperature(25).build();
+            dog = Dog.builder().nickname("Шарик").breed("Дворняга").build();
+            cat = Cat.builder().nickname("Мурзик").breed("Сиамская").build();
         }
         catch (Exception e){
             System.out.println(e.getMessage());
@@ -52,25 +45,30 @@ public class OneTableForHierarchyTestTable {
         session.getTransaction().commit();
         session.close();
     }
+
     @Test
     void save(){
+        session.persist(dog);
         session.persist(cat);
-        session.persist(fish);
     }
+
     @Test
-    void findFish(){
-       Fish fish1 = session.find(Fish.class,2);
-       System.out.println(fish1);
+    void findDog(){
+        Dog dog1 = session.find(Dog.class,1);
+        Assertions.assertEquals(dog1.getBreed(),"Дворняга");
     }
     @Test
     void findCat(){
-       Cat cat1 = session.find(Cat.class,1);
-       System.out.println(cat1);
+        Cat cat1 = session.find(Cat.class,2);
+        Assertions.assertEquals(cat1.getNickname(),"Мурзик");
     }
     @Test
-    void findAllPets(){
-        List<Pet> pets = session.createQuery("from Pet", Pet.class).getResultList();
-        pets.forEach(System.out::println);
+    void findAllAnimal(){
 
+        /*
+        Данный select будет использовать union all для получения данных из таблиц dog и cat.
+         */
+     List<Animal> animals = session.createQuery("SELECT bd FROM Animal bd", Animal.class).getResultList();
+     animals.forEach(System.out::println);
     }
 }
